% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{aldg}
\alias{aldg}
\title{Function to compute the aLDG measure}
\usage{
aldg(
  x,
  y,
  sx = NULL,
  sy = NULL,
  thred = -Inf,
  hx = NULL,
  hy = NULL,
  band = "fix",
  wd = 0.5,
  qd = 0.1,
  opt = TRUE,
  stat = "normgap",
  cutoff = 1,
  chooset = NULL
)
}
\arguments{
\item{hx, hy:}{(aLDG specific parameters) vector consist provided bandwidth for x (or y), default to NULL}

\item{x:}{a numeric vector}

\item{y:}{a numeric vector of the same length as x}

\item{sx:}{a list containing the sorting information, defualt to NULL,
if not NULL, then should have the format
sx = list(rank = vector of value rank, dist = matrix of pairwise distance)}

\item{sy:}{similar as the definition for sx, but for vector y}

\item{thred:}{(aLDG specific parameters) only the data > thred will be used for aLDG}

\item{band:}{(aLDG specific parameters) which method to use for automatically calculating bandwidth (if hx and hy are NULL)
-- 'fix': the window around x is \eqn{B_x = (x-wd \cdot h_n, x + wd \cdot h_n}
-- 'ada': the adaptive window around x is \eqn{P(X \in B_x \mid Y=y) = qd}}

\item{wd:}{(aLDG specific parameters) the coefficient before the bandwidth h = wd*h_n}

\item{qd:}{(aLDG specific parameters) the quantile window to compute the adaptive bandwidth \eqn{h_x = sd(x(qd))}}

\item{stat:}{(aLDG specific parameters) what statistics to use for \eqn{T_i} in aLDG}

\item{cutoff:}{(aLDG specific parameters) the coef c in threshold choice for aLDG:
\deqn{aLDG = \frac{1}{n} \sum_{i=1}^n ( T_i > \frac{\Phi^{-1}(1-n^{-c})}{n^{1/3}})}}
}
\value{
a vector
}
\description{
Function to compute the aLDG measure
}
\examples{
# Load example data
ans = aldg(runif(10),runif(10))
print(ans$aldg)
}
