% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{bidep}
\alias{bidep}
\title{Function to compute the bivariate dependency}
\usage{
bidep(
  x,
  y,
  sx = NULL,
  sy = NULL,
  methods = NULL,
  all = FALSE,
  thred = -Inf,
  hx = NULL,
  hy = NULL,
  band = "fix",
  wd = 1,
  qd = 0.1,
  opt = TRUE,
  stat = "normgap",
  cutoff = 1
)
}
\arguments{
\item{hx, hy:}{(aLDG specific parameters) vector consist provided bandwidth for x (or y), default to NULL}

\item{x:}{a numeric vector}

\item{y:}{a numeric vector of the same length as x}

\item{sx:}{a list containing the sorting information, defualt to NULL,
if not NULL, then should have the format
sx = list(rank = vector of value rank, dist = matrix of pairwise distance)}

\item{sy:}{similar as the definition for sx, but for vector y}

\item{methods:}{a vector of strings, should be set of the following:
-- 'Pearson': Pearaon's correlation
-- 'Spearman': Spearman's correaltion
-- 'Kendall': Kendall's \eqn{\tau}
-- 'TauStar': Kendall's \eqn{\tau^\star}
-- 'dCor': distance correlation
-- 'HSIC': Hilbert-Schmidt Independence Criterion (normalized to 0,1)
-- 'HoeffD': Hoeffding's D
-- 'HHG': HHG
-- 'MRank': Matching Ranks
-- 'MIC': Maximal Information Coefficient
-- 'aLDG': averaged Local Density Gap}

\item{all:}{logical value, if true then compute for all methods listed}

\item{thred:}{(aLDG specific parameters) only the data > thred will be used for aLDG}

\item{band:}{(aLDG specific parameters) which method to use for automatically calculating bandwidth (if hx and hy are NULL)
-- 'fix': the window around x is \eqn{B_x = (x-wd \cdot h_n, x + wd \cdot h_n}
-- 'ada': the adaptive window around x is \eqn{P(X \in B_x \mid Y=y) = qd}}

\item{wd:}{(aLDG specific parameters) the coefficient before the bandwidth h = wd*h_n}

\item{qd:}{(aLDG specific parameters) the quantile window to compute the adaptive bandwidth \eqn{h_x = sd(x(qd))}}

\item{opt:}{(aLDG specific parameters) for 'fix' bandwidth calculation,
if TRUE, we use the theoretical optimal rate \eqn{h_n = sd(x) n^{-1/6}};
if FALSE, we use \eqn{h_n = sd}}

\item{stat:}{(aLDG specific parameters) what statistics to use for \eqn{T_i} in aLDG}

\item{cutoff:}{(aLDG specific parameters) the coef c in threshold choice for aLDG:
\deqn{aLDG = \frac{1}{n} \sum_{i=1}^n ( T_i > \frac{\Phi^{-1}(1-n^{-c})}{n^{1/3}})}}
}
\value{
a vector
}
\description{
Estimate simulation parameters for the esco simulation from a real
dataset. See the individual estimation functions for more details on how this
is done.
}
\examples{
# Load example data
ans = bidep(runif(10),runif(10),methods='aLDG')
}
