% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{matdep}
\alias{matdep}
\title{Function to compute the pairwise dependency matrix for multivariate data}
\usage{
matdep(
  data,
  methods = NULL,
  all = FALSE,
  thred = -Inf,
  cutoff = 1,
  wd = 0.5,
  qd = 0.1,
  opt = TRUE,
  band = "fix",
  stat = "normgap",
  trial = 0,
  ncores = NULL,
  norm = FALSE,
  abs = FALSE,
  info = "",
  dir = "./dat/"
)
}
\arguments{
\item{thred, cutoff, wd, qd, opt, band, stat:}{same as that in function bidep, parameters specific to aLDG method}

\item{data:}{feature (row) by sample (column) matrix}

\item{methods:}{same as that in function bidep}

\item{all:}{same as that in function bidep, but omit 'hhg' since it takes too long}

\item{trial:}{integer denote the trial number}

\item{ncores:}{number of cores to use}

\item{norm:}{whether scale the value to \eqn{[0,1]}}

\item{abs:}{whether take absolute value of the value}

\item{info:}{information related to the settings}

\item{dir:}{the name of directory to save the results}
}
\value{
a list of matrix
}
\description{
Function to compute the pairwise dependency matrix for multivariate data
}
\examples{
Load example data
ans = matdep(runif(100,5,20), methods='aldg')
}
