% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simubi}
\alias{simubi}
\title{Function to simulate different bivariate dependency}
\usage{
simubi(n = 300, type = "linear", eps = 0.3)
}
\arguments{
\item{n:}{a integer specify the number of samples to be simulated}

\item{type:}{a character specify the type of dependency}

\item{eps:}{the noise level}
}
\value{
a list of two vectors
}
\description{
Function to simulate different bivariate dependency
}
\examples{
# Load example data
ans = simubi(100,'linear',0.3)
plot(ans$x, ans$y)
}
